requirements.txt
fastapi[all]==0.95.2
sqlalchemy==1.4.49
pydantic==1.10.12
python-dotenv==1.0.1
pyodbc==5.2.0